pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "slide-automation"
version = "0.1.0"
description = "Render structured JSON deck specs to PowerPoint using python-pptx and Sloan templates"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-pptx>=1.0.0",
]

[project.optional-dependencies]
mcp-server = [
    "mcp>=1.6.0,<2",
]

[project.scripts]
slide-render = "slide_automation.main:main"
slide-validate = "slide_automation.validate_cli:main"
slide-build = "slide_automation.build_cli:main"
slide-automation-mcp = "slide_automation.mcp.server_stdio:main"

[tool.setuptools.packages.find]
where = ["src"]
